+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.Blazor/ExampleJsInterop.cs
@@ -1,4 +1,4 @@
-using Microsoft.JSInterop;
+﻿using Microsoft.JSInterop;
 
 namespace ScottPlot.Blazor
 {