+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## ScottPlot 5.0.9-beta (in development)
 * Signal: Improved support for datasets with repeating values (#2933, #2935) _Thanks @StendProg_
+* Blazor: Added a Blazor control (#2959) _Thanks @sulivanganter_
 
 ## ScottPlot 4.1.68 (in development)
 * Axis: Added `IsReverse` property to let users invert the orientation of an axis (#2958) _Thanks @HandsomeGoldenKnight_